Cantor and dimension concepts.
